+++ core/goflow/engine_test.go
@@ -31,7 +31,7 @@ func TestEngineWebhook(t *testing.T) {
 	call, err := svc.Call(request)
 	assert.NoError(t, err)
 	assert.NotNil(t, call)
-	assert.Equal(t, "GET / HTTP/1.1\r\nHost: rapidpro.io\r\nUser-Agent: RapidProMailroom/Dev\r\nX-Mailroom-Mode: normal\r\nAccept-Encoding: gzip\r\n\r\n", string(call.RequestTrace))
+	assert.Equal(t, "GET / HTTP/1.1\r\nHost: rapidpro.io\r\nUser-Agent: Mailroom/Dev\r\nX-Mailroom-Mode: normal\r\nAccept-Encoding: gzip\r\n\r\n", string(call.RequestTrace))
 	assert.Equal(t, "HTTP/1.0 200 OK\r\nContent-Length: 2\r\n\r\n", string(call.ResponseTrace))
 	assert.Equal(t, "OK", string(call.ResponseBody))
 }
@@ -44,14 +44,15 @@ func TestSimulatorAirtime(t *testing.T) {
 
 	amounts := map[string]decimal.Decimal{"USD": decimal.RequireFromString(`1.50`)}
 
-	transfer, err := svc.Transfer(ctx, urns.URN("tel:+593979111111"), urns.URN("tel:+593979222222"), amounts, nil)
+	transfer, err := svc.Create(ctx, flows.NewEventUUID(), urns.URN("tel:+593979111111"), urns.URN("tel:+593979222222"), amounts, nil)
 	assert.NoError(t, err)
 
 	assert.Equal(t, &flows.AirtimeTransfer{
-		Sender:    urns.URN("tel:+593979111111"),
-		Recipient: urns.URN("tel:+593979222222"),
-		Currency:  "USD",
-		Amount:    decimal.RequireFromString(`1.50`),
+		ExternalID: "123456789",
+		Sender:     urns.URN("tel:+593979111111"),
+		Recipient:  urns.URN("tel:+593979222222"),
+		Currency:   "USD",
+		Amount:     decimal.RequireFromString(`1.50`),
 	}, transfer)
 }
 
@@ -72,7 +73,7 @@ func TestSimulatorWebhook(t *testing.T) {
 	call, err := svc.Call(request)
 	assert.NoError(t, err)
 	assert.NotNil(t, call)
-	assert.Equal(t, "GET / HTTP/1.1\r\nHost: rapidpro.io\r\nUser-Agent: RapidProMailroom/Dev\r\nX-Mailroom-Mode: simulation\r\nAccept-Encoding: gzip\r\n\r\n", string(call.RequestTrace))
+	assert.Equal(t, "GET / HTTP/1.1\r\nHost: rapidpro.io\r\nUser-Agent: Mailroom/Dev\r\nX-Mailroom-Mode: simulation\r\nAccept-Encoding: gzip\r\n\r\n", string(call.RequestTrace))
 	assert.Equal(t, "HTTP/1.0 200 OK\r\nContent-Length: 2\r\n\r\n", string(call.ResponseTrace))
 	assert.Equal(t, "OK", string(call.ResponseBody))
 }